:recycle: (Login) Rename schema file
